do not install ESP on zfs disks, use sgdisk instead of parted
ZFS has this use_whole_disk concept, so we try to partition the same
way as zfs does by default. There is room at start of disk to insert
a grub boot partition. But adding a EFI ESP is not possible.

Note: zfs people think this is just a waste of space an not
required.  Instead, you should put the ESP on another disk.

Turns out that parted is unable to set partition numbers, so
we use sgdisk instead.