hd_list: skip all devices with iso9660 as FS
This ensures that when booting from a USB stick, or similar device,
we do not show that device in the "Target Harddisk" selector.

While there may be some constructed cases where this actually worked
and was wanted, e.g., one created a partition on the end of a USB
storage device, put the ISO on that and used the beginning of that
device as target, I'd say that even if that worked with luck, it's
better not actively not support it.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>