avoid doing package specific cleanups installer
Packages which we have under control have to ensure themself that the
correct permissions are set, else the "install on top of Debian" also
fails. It just crowd proxinstall unecessarily, so avoid to add new
ones if possible. This specific one is handled in the respective
package since quite a bit.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>