set console keymap on language selection
This is an enhancement for #2804

Having a shell waiting on vt3, it improves user experience if it has the
same keyboard map as the X11 installer.

This is accomplished by setting the contents of '/etc/default/keyboard' and
then running `setupcon`. Simply calling `loadkeys` would not work, since the
keymaps in debian are generated from the x11 definitions by ckbcomp and then
saved in '/etc/console-setup/'.

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>