fix 2925: edit /etc/kernel/cmdline unconditionally
currently the installer prepares the ESP disks for ZFS, irrespective of the
boot-mode (EFI, legacy) - in order to enable users to change the boot-mode
in the BIOS and keep the system bootable.

This patch updates /etc/kernel/cmdline in both boot-modes, which is necessary
to make the system actually bootable (else the systemd-boot config uses
the cmdline from the installer).

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>