override grub default config via installer
We used to ship our own grub in the installer, originally due to ZFS
support, but we also used it to setup a slightly modified
`/etc/default/grub`. We do not need a self-build grub for technical
reasons like ZFS support anymore, so avoid the need to ship it and
move over deploying the config changes to the installer.

Just write out the whole config, much easier to see what's going on
compared to some `sed` calls.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>