unconfigured: heuristically detect high-dpi and adapt font/GDK_SCALE
Use the biggest terminus font we can get our hands on, and use the
Uni2 variant, which seems like an OK trade-off (the installer is
rather ascii anyway).

The detection uses the console columns in such away that high-dpi
screen should be detected but just wide-formats (like 21:9) should
not. This is a bit brittle, but even miss-detecting it should still
allow using it (the installer is made for 1024x768, which when
doubled still fits on WQXGA (2560x1600) and mostly on 1440p too.

It could be more robust to depend on edid-decode, but as physical
dimensions are not required there, and even if specified often quite
off, it's not a sure thing either. So go for simple first.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>