get_device_desc: display disk size with 2 decimal places
instead of cutting it off by using int

especially relevant currently with 2 TB disks (people wonder why
it's displayed as 1 TiB - and would be fine with 1.95TiB).

Reported in our community forum:
https://forum.proxmox.com/threads/.103636/

Tested in a VM with a 100 GiB, 20GiB and 2TiB disk

Addtionally added the explicit return ath the end of the if branch,
instead of relying on the last statement not moving.

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>