fix #3188: update hdsize spin-button on disk-selection change
the patch splits get_hdsize_spinbtn into a sub for generating the adjustment
gtk-element and one for the button itself and moves them above
create_raid_disk_raid (so that the adjustment sub can be called)

The adjustment gets it's data from the disks selected for installation
in the GUI, if a size is not provided.

Additionally the bogus 'zfs' argument to the create_raid_advanced_grid
was dropped.

Tested with a VM with 2x 100G disks, and one with 20G

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>