disks: raid: avoid restting hdsize when reopening options
The"auto get lowest disk size from selected set" logic also reset it
to that if the previously chosen value was already lower on reopening
the disk dialogue.

That can be rather surprising for an user as the hdsize field is a
bit hidden in the advanced dialogue, so just reopening to ensure the
correct disks are selected may interfere with a previously lowered
hdsize.

Reported-by: Fabian Ebner <f.ebner@proxmox.com>
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>