drop enabling timesyncd, wrong since chrony switch & unnecessary anyway
It should actually fail since the switch to chrony, but syscmd only
returns the exit code and we did not check it.

Shortly thought about checking for the existence of
"$targetdir/lib/systemd/system/chrony.service" and enable
chrony.service in that case, but that is unnecessary too, as we know
due to this bogus enable timesyncd which never stopped chrony to run
happily enabled after installation anyway.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>