create lvm: improve rounding/alignment
stay in MB size range for longer to have some accuracy level we need
and mask the last 2 bits (= 4 MB) to align down to a level that
should not interfere with LVM anymore.

Otherwise, with $rest sizes that hit the second if branch < 48
regressed on LVM due to missing one extend of expected space.

Dominik verified this off-list to fix the case Stefan run into.

Reported-by: Stefan Hanreich <s.hanreich@proxmox.com>
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>