create lvm: align LV size down to 4MB boundary for non-pve products
Following commit 1a5fa7b0936381314f613e525255078093f7f258 but for
the non-PVE code path we also need to align down to 4 MB with the
adapted auto-sizing, especially on smaller disks.

For PMG and PBS installations. Without it installation fails due to
one missing extent when creating the root LV. Minimally tested on a
VM of mine

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>
 [ T: reworded commit message a bit ]
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>